signaloneof: set signal per process
Similar to flow control, allow setting the termination signal sent when
the control process shuts down for each child process.

The alcove control process setting for signaloneof now sets the default
on fork().

I'm not sure how useful fine grained control of the signal is:
it might be useful if you have a long running job that you don't want
interrupted:

    {ok, Child} = alcove:fork(Drv, []),
    % insert real, long running job here, it should ignore EOF of stdin
    ok = alcove:execvp(Drv, [Child], "sleep", ["run-for-an-hour", "3600"]),
    alcove:setopt(Drv, [], signaloneof, Child, 0).

Previously the same thing could be done for all child processes of the
control process.

+++ c_src/alcove_event_loop.c
@@ -750,8 +750,9 @@ signal_pid(alcove_state_t *ap, alcove_child_t *c, void *arg1, void *arg2)
     UNUSED(arg2);
 
     if (c->fdctl == ALCOVE_CHILD_EXEC) {
-      (void)kill(c->pid, ap->signaloneof);
-      (void)kill(c->pid, SIGCONT);
+      if (kill(c->pid, c->signaloneof) == 0) {
+        (void)kill(c->pid, SIGCONT);
+      }
     }
 
     return 1;
